requirements.txt
dominate==2.4.0
runway-python==0.5.3
scipy==1.3.1
torch==0.3.1
torchvision==0.2.0
visdom==0.1.8.9
